+++ src/main/java/nc/vista/cliente/ItemMenuPanelCliente.java
@@ -4,6 +4,7 @@
  */
 package nc.vista.cliente;
 
+import nc.vista.vendedor.dialog.ItemMenuVer;
 import nc.vista.vendedor.*;
 import java.util.Arrays;
 import java.util.List;
