+++ src/main/java/nc/vista/cliente/dialog/ItemMenuVerCliente.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JDialog.java to edit this template
  */
-package nc.vista.cliente;
+package nc.vista.cliente.dialog;
 
 import nc.vista.vendedor.*;
 import java.util.ArrayList;
