+++ src/main/java/nc/vista/initView/ClientesMenuPanel.java
@@ -4,6 +4,8 @@
  */
 package nc.vista.initView;
 
+import nc.vista.initView.dialog.DialogCliente;
+import nc.vista.initView.dialog.CreacionCliente;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -14,8 +16,7 @@ import javax.swing.table.TableRowSorter;
 import nc.controlador.ClientController;
 import nc.vista.PersonalizatedTableModel;
 import nc.vista.VistaPrincipal;
-import nc.vista.cliente.CreacionCliente;
-import nc.vista.cliente.ModificarCliente;
+import nc.vista.initView.dialog.ModificarCliente;
 
 /**
  *
@@ -69,7 +70,7 @@ public class ClientesMenuPanel extends javax.swing.JPanel {
         clients.modificarCliente(id, name, dir, mail, id, coorX, coorY);
         updateModel();
     }
-     public void crearModificarCliente(int id, String name, String dir, String mail, String cuit, double coorX, double coorY){
+    public void crearModificarCliente(int id, String name, String dir, String mail, String cuit, double coorX, double coorY){
          ModificarCliente modificarCliente = new ModificarCliente(id, name, dir, mail, cuit, coorX, coorY, this);
          modificarCliente.setVisible(true);
          modificarCliente.setLocationRelativeTo(null);
