+++ src/main/java/nc/vista/initView/dialog/ModificarCliente.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
  */
-package nc.vista.cliente;
+package nc.vista.initView.dialog;
 
 import nc.vista.initView.ClientesMenuPanel;
